REFACTOR: Changed the project structure to support inner-folders
