+++ Makefile
@@ -9,13 +9,15 @@ INCDIR := headers
 BUILDDIR := build
 
 
-SOURCES := $(wildcard $(SRCDIR)/*.cpp)
+SRCDIRS := $(shell find $(SRCDIR) -type d)
+SOURCES := $(foreach dir,$(SRCDIRS),$(wildcard $(dir)/*.cpp))
+
 OBJECTS := $(patsubst $(SRCDIR)/%.cpp,$(BUILDDIR)/%.o,$(SOURCES))
 MAIN := main.cpp
 
 
 TARGET := my_game
-RM := rm -f
+RM := rm -rf
 
 ifeq ($(OS),Windows_NT)
     # Change forward slashes to backslashes for Windows paths
@@ -26,13 +28,15 @@ ifeq ($(OS),Windows_NT)
 endif
 
 
-$(TARGET): $(OBJECTS) $(MAIN)
+$(TARGET):  $(OBJECTS) $(MAIN) 
 	$(CC) $(CFLAGS) -I$(INCDIR) -o $@ $(MAIN) $(OBJECTS) $(SFML_LIBS) -g 
 
+BUILDDIRS:
+	mkdir -p $(BUILDDIR)
 
-$(BUILDDIR)/%.o: $(SRCDIR)/%.cpp
+$(BUILDDIR)/%.o: $(SRCDIR)/%.cpp | BUILDDIRS
+	mkdir -p $(dir $@)
 	$(CC) $(CFLAGS) -I$(INCDIR) -c -o $@ $< -g
 
-
 clean:
-	$(RM) $(BUILDDIR)*.o $(TARGET)
+	$(RM) $(BUILDDIR) $(TARGET)
