+++ headers/screens/PlaySettings.hpp
@@ -5,8 +5,5 @@
 
 void playSettingsDraw(sf::RenderWindow &window, Menu &psMenu);
 void playSettingsUpdate(sf::RenderWindow &window, Menu &psManu);
-void numberOfPlayersUpdate(sf::RenderWindow &window);
-void numberOfPlayersDraw(sf::RenderWindow &window);
 Menu initlizePlaySettingsMenu();
-Menu initlizeNumberOfPlayersMenu();
 #endif // PLYASETTINGS_H
